Make Stages officially public and sealed.
