+++ src/internal/system/schedule/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod raw_task;
 pub(crate) mod sendable;
+pub(crate) mod stage;
 pub(crate) mod task;
